moved tests to other_test and removed upgrade test for example
