+++ tests/pr_test.go
@@ -10,21 +10,13 @@ import (
 
 	"github.com/gruntwork-io/terratest/modules/files"
 	"github.com/gruntwork-io/terratest/modules/logger"
-	"github.com/gruntwork-io/terratest/modules/random"
 	"github.com/gruntwork-io/terratest/modules/terraform"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/common"
-	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testhelper"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testschematic"
 )
 
-// Use existing resource group
-const resourceGroup = "geretain-test-resources"
-
-// Ensure every example directory has a corresponding test
-const advancedExampleDir = "examples/advanced"
-const basicExampleDir = "examples/basic"
 const configurableDADir = "solutions/fully-configurable"
 
 var validRegions = []string{
@@ -37,91 +29,6 @@ var validRegions = []string{
 	"jp-tok",
 	"us-south",
 	"us-east",
-}
-
-func setupOptions(t *testing.T, prefix string, dir string) *testhelper.TestOptions {
-	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
-		Testing:       t,
-		TerraformDir:  dir,
-		Prefix:        prefix,
-		ResourceGroup: resourceGroup,
-		Region:        validRegions[rand.Intn(len(validRegions))],
-	})
-	return options
-}
-
-// Consistency test for the basic example
-func TestRunBasicExample(t *testing.T) {
-	t.Parallel()
-
-	options := setupOptions(t, "icm-basic", basicExampleDir)
-
-	output, err := options.RunTestConsistency()
-	assert.Nil(t, err, "This should not have errored")
-	assert.NotNil(t, output, "Expected some output")
-}
-
-func TestRunAdvancedExampleInSchematics(t *testing.T) {
-	t.Parallel()
-
-	var region = validRegions[rand.Intn(len(validRegions))]
-
-	options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
-		Testing: t,
-		Prefix:  "icm-adv",
-		TarIncludePatterns: []string{
-			"*.tf",
-			"modules/metrics_routing" + "/*.tf",
-			advancedExampleDir + "/*.tf",
-		},
-		ResourceGroup:          resourceGroup,
-		TemplateFolder:         advancedExampleDir,
-		Tags:                   []string{"test-schematic"},
-		DeleteWorkspaceOnFail:  false,
-		WaitJobCompleteMinutes: 60,
-	})
-
-	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-		{Name: "prefix", Value: options.Prefix, DataType: "string"},
-		{Name: "region", Value: region, DataType: "string"},
-	}
-
-	err := options.RunSchematicTest()
-	assert.Nil(t, err, "This should not have errored")
-}
-
-// Upgrade test in schematics (using advanced example)
-func TestRunUpgradeExampleInSchematics(t *testing.T) {
-	t.Parallel()
-
-	var region = validRegions[rand.Intn(len(validRegions))]
-
-	options := testschematic.TestSchematicOptionsDefault(&testschematic.TestSchematicOptions{
-		Testing: t,
-		Prefix:  "icm-adv-upg",
-		TarIncludePatterns: []string{
-			"*.tf",
-			"modules/metrics_routing" + "/*.tf",
-			advancedExampleDir + "/*.tf",
-		},
-		ResourceGroup:          resourceGroup,
-		TemplateFolder:         advancedExampleDir,
-		Tags:                   []string{"test-schematic"},
-		DeleteWorkspaceOnFail:  false,
-		WaitJobCompleteMinutes: 60,
-	})
-
-	options.TerraformVars = []testschematic.TestSchematicTerraformVar{
-		{Name: "ibmcloud_api_key", Value: options.RequiredEnvironmentVars["TF_VAR_ibmcloud_api_key"], DataType: "string", Secure: true},
-		{Name: "prefix", Value: options.Prefix, DataType: "string"},
-		{Name: "region", Value: region, DataType: "string"},
-	}
-
-	err := options.RunSchematicUpgradeTest()
-	if !options.UpgradeTestSkipped {
-		assert.Nil(t, err, "This should not have errored")
-	}
 }
 
 func TestRunFullyConfigurable(t *testing.T) {
@@ -167,7 +74,7 @@ func TestRunUpgradeFullyConfigurable(t *testing.T) {
 	t.Parallel()
 
 	var region = validRegions[rand.Intn(len(validRegions))]
-	prefix := fmt.Sprintf("icm-da-up-%s", strings.ToLower(random.UniqueId()))
+	prefix := "icm-da-up-"
 
 	// ------------------------------------------------------------------------------------
 	// Provision Cloud Monitoring
